+++ BoardBall_Game/Engine.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <Windows.h>
+
+void Init();
+void Draw_Frame(HDC hdc);
